# contracts/openapi/tool-defect-api-v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "刀具缺陷检测系统接口",
    "version": "1.0.0",
    "description": "采集、业务查询、复核、数据集、训练、模型发布和内部推理接口的 v1 唯一契约源。"
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "地址由环境配置提供，示例域名不可用于生产"
    }
  ],
...
  "paths": {
...
    "/api/v1/images/{image_id}/access-ticket": {
      "post": {
        "operationId": "createImageAccessTicket",
        "tags": [
          "detections"
        ],
        "security": [
          {
            "UserOidc": [
              "image:view"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ImageId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageAccessTicketRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "短时只读访问授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageAccessTicketResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "createImageAccessTicket"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "EdgeDeviceMtls": {
        "type": "mutualTLS"
      },
      "EdgeOAuth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example.invalid/oauth2/token",
            "scopes": {
              "capture:write": "写入采集事件",
              "capture:read": "读取采集状态",
              "device:heartbeat": "上报设备心跳"
            }
          }
        }
      },
      "UserOidc": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://identity.example.invalid/.well-known/openid-configuration"
      },
      "ServiceMtls": {
        "type": "mutualTLS"
      },
      "ServiceOAuth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example.invalid/oauth2/token",
            "scopes": {
              "inference:callback": "提交推理执行结果",
              "runtime:read": "读取推理运行时健康"
            }
          }
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 256
        }
      },
...
      "ImageId": {
        "name": "image_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "请求格式错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "身份认证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Forbidden": {
        "description": "没有资源或操作权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "幂等摘要或乐观锁冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "ValidationError": {
        "description": "语义或状态校验失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ImageAccessTicketRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "purpose"
        ],
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "VIEW",
              "DOWNLOAD"
            ]
          }
        }
      },
      "ImageAccessTicketResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "method",
          "url",
          "expires_at"
        ],
        "properties": {
          "method": {
            "const": "GET"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "pattern": "^https://"
          },
          "expires_at": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
          }
        }
      },
...
    }
  },
...
}